refactor: consolidate notion clients
