+++ cvbuilder/notion/__init__.py
@@ -1,0 +1,19 @@
+"""Notion integration package."""
+
+from .projects import Projects
+from .personal import PersonalInfo
+from .experience import Experience
+from .certificates import Certificates
+from .education import Education
+from .skills import generate_skills_from_projects
+from .notion import Notion
+
+__all__ = [
+    "Projects",
+    "PersonalInfo",
+    "Experience",
+    "Certificates",
+    "Education",
+    "generate_skills_from_projects",
+    "Notion",
+]
